+++ .gitignore
@@ -1,4 +1,1 @@
-.idea
-.bigquery
-.repl
 target
