+++ src/main/scala/com/snowplowanalytics/snowplow/storage/bigquery/loader/Main.scala
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) 2018 Snowplow Analytics Ltd. All rights reserved.
+ *
+ * This program is licensed to you under the Apache License Version 2.0,
+ * and you may not use this file except in compliance with the Apache License Version 2.0.
+ * You may obtain a copy of the Apache License Version 2.0 at http://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the Apache License Version 2.0 is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
+ */
+package com.snowplowanalytics.snowplow.storage.bigquery.loader
+
+import com.spotify.scio._
+
+object Main {
+  def main(cmdlineArgs: Array[String]): Unit = {
+    val (sc, args) = ContextAndArgs(cmdlineArgs)
+    val env = CommandLine.parse(args)
+    EventsTable.create(env.config.projectId, env.config.datasetId, env.config.tableId)
+    // Loader.run(config, sc)
+    // val _ = sc.close().waitUntilDone()
+  }
+}
