--- src/main/scala/com/snowplowanalytics/snowplow/storage/bqloader/CommandLine.scala
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2018 Snowplow Analytics Ltd. All rights reserved.
- *
- * This program is licensed to you under the Apache License Version 2.0,
- * and you may not use this file except in compliance with the Apache License Version 2.0.
- * You may obtain a copy of the Apache License Version 2.0 at http://www.apache.org/licenses/LICENSE-2.0.
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the Apache License Version 2.0 is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
- */
-package com.snowplowanalytics.snowplow.storage.bqloader
-
-import cats.syntax.either._
-
-import com.spotify.scio.Args
-
-import core.Config.{ transform, decodeBase64Json, Environment, EnvironmentConfig }
-
-/**
-  * Loader specific CLI configuration
-  * Unlike Mutator, required --key=value format and ignores unknown options (for Dataflow)
-  */
-object CommandLine {
-  def parse(args: Args): Environment = {
-    val environment =
-      for {
-        c <- decodeBase64Json(args("config"))
-        r <- decodeBase64Json(args("resolver"))
-        e <- transform(EnvironmentConfig(r, c))
-      } yield e
-
-    environment.fold(throw _, id)
-  }
-
-  private def id[A](a: A): A = a
-}
